Add main menu and UUIDv4 functions
